+++ src/modules/cpu/option.h
@@ -10,4 +10,5 @@ typedef struct FFCPUOptions
     FFModuleArgs moduleArgs;
 
     bool temp;
+    uint8_t freqNdigits;
 } FFCPUOptions;